Conflicts:
	db/version_set.cc
	include/leveldb/options.h
	util/options.cc

+++ db/db_impl.cc
@@ -10,6 +10,7 @@
 #include <set>
 #include <string>
 #include <stdint.h>
+#include <stdexcept>
 #include <vector>
 #include <unordered_set>
 
@@ -2015,6 +2016,16 @@ int64_t DBImpl::TEST_MaxNextLevelOverlappingBytes() {
 Status DBImpl::Get(const ReadOptions& options,
                    const Slice& key,
                    std::string* value) {
+  return GetImpl(options, key, value);
+}
+
+// If no_IO is true, then returns Status::NotFound if key is not in memtable,
+// immutable-memtable and bloom-filters can guarantee that key is not in db,
+// "value" is garbage string if no_IO is true
+Status DBImpl::GetImpl(const ReadOptions& options,
+                       const Slice& key,
+                       std::string* value,
+                       const bool no_IO) {
   Status s;
 
   StopWatch sw(env_, options_.statistics, DB_GET);
@@ -2043,12 +2054,12 @@ Status DBImpl::Get(const ReadOptions& options,
   // s is both in/out. When in, s could either be OK or MergeInProgress.
   // value will contain the current merge operand in the latter case.
   LookupKey lkey(key, snapshot);
-  if (mem->Get(lkey, value, &s, options_)) {
+  if (mem->Get(lkey, value, &s, options_, no_IO)) {
     // Done
-  } else if (imm.Get(lkey, value, &s, options_)) {
+  } else if (imm.Get(lkey, value, &s, options_, no_IO)) {
     // Done
   } else {
-    current->Get(options, lkey, value, &s, &stats, options_);
+    current->Get(options, lkey, value, &s, &stats, options_, no_IO);
     have_stat_update = true;
   }
   mutex_.Lock();
@@ -2138,6 +2149,12 @@ std::vector<Status> DBImpl::MultiGet(const ReadOptions& options,
   return statList;
 }
 
+bool DBImpl::KeyMayExist(const Slice& key) {
+  std::string value;
+  const Status s = GetImpl(ReadOptions(), key, &value, true);
+  return !s.IsNotFound();
+}
+
 Iterator* DBImpl::NewIterator(const ReadOptions& options) {
   SequenceNumber latest_snapshot;
   Iterator* internal_iter = NewInternalIterator(options, &latest_snapshot);
@@ -2173,6 +2190,10 @@ Status DBImpl::Merge(const WriteOptions& o, const Slice& key,
 }
 
 Status DBImpl::Delete(const WriteOptions& options, const Slice& key) {
+  if (options_.deletes_check_filter_first && !KeyMayExist(key)) {
+    RecordTick(options_.statistics, NUMBER_FILTERED_DELETES);
+    return Status::OK();
+  }
   return DB::Delete(options, key);
 }
 
@@ -2232,13 +2253,19 @@ Status DBImpl::Write(const WriteOptions& options, WriteBatch* my_batch) {
       }
       if (status.ok()) {
         status = WriteBatchInternal::InsertInto(updates, mem_);
+        if (!status.ok()) {
+          // Panic for in-memory corruptions
+          // Note that existing logic was not sound. Any partial failure writing
+          // into the memtable would result in a state that some write ops might
+          // have succeeded in memtable but Status reports error for all writes.
+          throw std::runtime_error("In memory WriteBatch corruption!");
+        }
+        versions_->SetLastSequence(last_sequence);
+        last_flushed_sequence_ = current_sequence;
       }
       mutex_.Lock();
     }
-    last_flushed_sequence_ = current_sequence;
     if (updates == &tmp_batch_) tmp_batch_.Clear();
-
-    versions_->SetLastSequence(last_sequence);
   }
 
   while (true) {
